+++ scripts/launch_integration_tests.sh
@@ -1,18 +1,5 @@
 #!/bin/bash
+set -e
 
-# delete previous scenario data
-python -m mxops \
-            data \
-            delete \
-            -n DEV \
-            -s integration_tests
-
-# execute integrations scenes
-python -m mxops \
-            execute \
-            -n DEV \
-            -s integration_tests \
-            integration_tests/scenes/accounts/devnet_accounts.yaml \
-            integration_tests/scenes
-
-# check correct execution
+integration_tests/piggy_bank/scripts/run_test.sh
+integration_tests/wrapping/scripts/run_test.sh