+++ scripts/launch_tests.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-coverage run -m pytest
-coverage html